temp save (findRecord)

+++ util.h
@@ -8,7 +8,7 @@ class Util : public QObject
     Q_OBJECT
 public:
     static QString formatTime(int ms);
-    static QString findRecord(const QString &fileName, qint64 pos);
+    static QString findRecord(const QString &fileName, quint32 pos);
     static quint8 hash(const QString &s);
 };
